requirements.txt
arxiv
google-generativeai
python-dotenv
PyYAML
requests
aiohttp
aiofiles
reportlab
